pyproject.toml
[tool.poetry]
name = "sql_test"
version = "0.1.0"
description = ""
authors = ["tsugumi-sys <tidemark0105@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
SQLAlchemy = "^1.4.31"
PyMySQL = "^1.0.2"
google-api-python-client = "^2.36.0"
python-dotenv = "^0.19.2"
cloud-sql-python-connector = {extras = ["pymysql"], version = "^0.5.0"}
fastapi = "^0.72.0"
uvicorn = {extras = ["standard"], version = "^0.17.0"}
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
